Fetch entity

+++ model/db/variant.h
@@ -4,7 +4,7 @@
 #include <variant>
 
 namespace model::db {
-using Variant = std::variant<int, long long, std::string>;
+using Variant = std::variant<int32_t, int64_t, std::string>;
 }  // namespace model::db
 
 #endif // VARIANT_H
